rework css var replacement
